requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.3.0
